refactor: :recycle: Extracted the fetching of user portfolio to the UserAssetContext
