+++ frontend/src/Common/Contexts/UserBalanceContext/index.tsx
@@ -2,19 +2,23 @@ import { ReactChildren } from "Common/Types";
 import { createContext, useEffect, useState } from "react";
 import useFetchUserBalance from "./CustomHooks/useFetchUserBalance";
 import usePostDeposit from "./CustomHooks/usePostDeposit";
+import IStock from "Interfaces/IStock";
+import useFetchPortfolio from "./CustomHooks/fetchPortfolio";
 
 interface Props {
     children: ReactChildren
 }
 
-interface IUserBalanceContext {
+interface IUserAssetContext {
     userBalance: number
     postDeposit: (funds:number) => Promise<void>
+    stockList: IStock[]
 }
 
-const UserBalanceContext = createContext<IUserBalanceContext>(undefined!)
+const UserAssetContext = createContext<IUserAssetContext>(undefined!)
 
-const UserBalanceProvider = ({ children }: Props) => {
+const UserAssetProvider = ({ children }: Props) => {
+    //set state that will manage user balance
     const [userBalance, setUserBalance] = useState(0)
 
     //fetch currnet user balance from server and set state to the response
@@ -27,11 +31,23 @@ const UserBalanceProvider = ({ children }: Props) => {
     //function that post fund deposits to the server
     const postDeposit = usePostDeposit(setUserBalance)
 
+    //set state that will manage the user portfolio
+    const [stockList, setStockList] = useState<IStock[]>([])
+
+    //fetch the stock data from db and set state to the response
+    const fetchPortfolio = useFetchPortfolio()
+    useEffect(() => {
+        const response = fetchPortfolio()
+        response.then(portfolio => {
+            setStockList(portfolio!)
+        })
+    }, [fetchPortfolio])
+
     return (
-        <UserBalanceContext.Provider value={{ userBalance, postDeposit }}>
+        <UserAssetContext.Provider value={{ userBalance, postDeposit, stockList }}>
             {children}
-        </UserBalanceContext.Provider>
+        </UserAssetContext.Provider>
     )
 }
 
-export { UserBalanceContext, UserBalanceProvider }
+export { UserAssetContext, UserAssetProvider }
