+++ frontend/src/routes/AuthRequiredRoutes/AuthChecker/index.tsx
@@ -4,7 +4,7 @@ import { Outlet } from "react-router-dom";
 import { SessionContext } from "Common/Contexts/SessionContext";
 import { useContext, useEffect } from "react";
 import { SnackbarContext } from "Common/Contexts/SnackbarContext";
-import { UserBalanceProvider } from "Common/Contexts/UserBalanceContext";
+import { UserAssetProvider } from "Common/Contexts/UserBalanceContext";
 
 /**
  * Function that acts as middleware, checking whether user is already
@@ -29,9 +29,9 @@ function AuthRequestBranch() {
     
     return(
         loggedIn ? 
-            <UserBalanceProvider>
+            <UserAssetProvider>
                 <Outlet/>
-            </UserBalanceProvider> : 
+            </UserAssetProvider> : 
             <LoginPage/> 
     )
 }
